01-quick-start/python/pyproject.toml
[project]
name = "quick-start-python"
version = "0.1.0"
description = "Phonic Examples: Quick Start Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "phonic==0.30.7",
    "python-dotenv==1.1.1",
]

[project.optional-dependencies]
dev = [
    "black==25.9.0",
    "isort==6.0.1",
]
